Fee AMM Enumerate over usages
